refactor(assess): move feature-maturity logic into queryissue
The maturity-guidance logic (which derives an issue's description from its
maturity in the target version) was living in cmd/assessMigrationCommand.go even
though it is issue semantics. Relocate it next to the issues it acts on.

- Move CheckIssueSupportMaturityInTDBVersion, buildExperimentalMaturityAnnotation,
  buildNativeResolutionRecommendation, maturityTierName, and GetSupportedVersions
  (renamed from FormatSupportedVersions) into queryissue/parser_issue_detector.go.
  CheckIssueSupportMaturityInTDBVersion now takes the target version as a param
  instead of reading a cmd package global.
- Annotate the issue Description once, in the parser's single version-filter choke
  point (getIssuesNotFixedInTargetDbVersion), so every detection path (assessment,
  analyze-schema, query constructs) reports maturity consistently. The cmd
  converters and the query-construct build now just use the pre-annotated
  Description.
- Bind GetSupportedVersions directly to the report template funcMap (passing the
  three version maps) and drop the cmd getSupportedVersions wrapper.
- Relocate the maturity unit tests into parser_issue_detector_test.go; the shared
  parser comparison loops apply the same annotation to the expected issue so the
  full-struct comparison still holds.

Pure relocation: assessment and analyze-schema reports are byte-unchanged
(build/vet/unit green; assessment-report-test and analyze-schema verified).

---
_automated · Claude Code (Opus 4.8)_
