+++ yb-voyager/cmd/common_test.go
@@ -26,10 +26,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/yugabyte/yb-voyager/yb-voyager/src/constants"
-	"github.com/yugabyte/yb-voyager/yb-voyager/src/issue"
 	"github.com/yugabyte/yb-voyager/yb-voyager/src/migassessment"
-	"github.com/yugabyte/yb-voyager/yb-voyager/src/query/queryissue"
 	"github.com/yugabyte/yb-voyager/yb-voyager/src/utils"
 	"github.com/yugabyte/yb-voyager/yb-voyager/src/ybversion"
 	testutils "github.com/yugabyte/yb-voyager/yb-voyager/test/utils"
@@ -764,139 +761,4 @@ func Int64Ptr(i int64) *int64 {
 
 func StringPtr(s string) *string {
 	return &s
-}
-
-// ================== Feature maturity (Tech Preview / Early Access) ==============================
-
-func TestFormatSupportedVersions(t *testing.T) {
-	gaMap := map[string]*ybversion.YBVersion{ybversion.SERIES_2025_2: ybversion.V2025_2_0_0}
-	eaMap := map[string]*ybversion.YBVersion{ybversion.SERIES_2025_1: ybversion.V2025_1_0_0}
-	tpMap := map[string]*ybversion.YBVersion{ybversion.SERIES_2024_2: ybversion.V2024_2_0_0}
-
-	// All three tiers, deterministic (sorted) output. GA is untagged; TP/EA are tagged.
-	got := formatSupportedVersions(gaMap, eaMap, tpMap)
-	assert.Equal(t,
-		">=2024.2.0.0 (2024.2 series) (TP), >=2025.1.0.0 (2025.1 series) (EA), >=2025.2.0.0 (2025.2 series)",
-		got)
-
-	// Only an EA entry.
-	got = formatSupportedVersions(nil, eaMap, nil)
-	assert.Equal(t, ">=2025.1.0.0 (2025.1 series) (EA)", got)
-
-	// All empty -> empty string.
-	assert.Equal(t, "", formatSupportedVersions(nil, nil, nil))
-}
-
-func TestBuildExperimentalMaturityAnnotation(t *testing.T) {
-	v := ybversion.V2025_2_0_0
-
-	// TP, with flags.
-	assert.Equal(t,
-		fmt.Sprintf("This feature is available as Tech Preview (TP) in the target version (2025.2.0.0) — %s, and is not enabled by default. Enable with the flag(s): yb_enable_foo=true, yb_bar=64.", constants.TP_MATURITY_CAVEAT),
-		buildExperimentalMaturityAnnotation(constants.MATURITY_TP, v, []string{"yb_enable_foo=true", "yb_bar=64"}))
-
-	// EA, no flags -> no "Enable with" sentence.
-	assert.Equal(t,
-		fmt.Sprintf("This feature is available as Early Access (EA) in the target version (2025.2.0.0) — %s, and is not enabled by default.", constants.EA_MATURITY_CAVEAT),
-		buildExperimentalMaturityAnnotation(constants.MATURITY_EA, v, nil))
-
-	// Non-experimental maturities produce no annotation.
-	assert.Empty(t, buildExperimentalMaturityAnnotation(constants.MATURITY_GA, v, nil))
-	assert.Empty(t, buildExperimentalMaturityAnnotation(constants.MATURITY_UNSUPPORTED, v, nil))
-}
-
-func TestBuildNativeResolutionRecommendation(t *testing.T) {
-	flags := []string{"yb_enable_derived_saops=true", "yb_max_saop_merge_streams=64"}
-
-	// Resolution is TP in the target version.
-	assert.Equal(t,
-		fmt.Sprintf("Consider using bucket-based indexes — available as Tech Preview (TP) in the target version (2025.2.1.0) — %s, and is not enabled by default. Enable with the flag(s): yb_enable_derived_saops=true, yb_max_saop_merge_streams=64.", constants.TP_MATURITY_CAVEAT),
-		buildNativeResolutionRecommendation("bucket-based indexes", constants.MATURITY_TP, ybversion.V2025_2_1_0, "", flags))
-
-	// Resolution not in target yet, but available later (UNSUPPORTED maturity + supportedVersions).
-	assert.Equal(t,
-		"Consider using bucket-based indexes — available in >=2025.2.1.0 (2025.2 series) (TP). Enable with the flag(s): yb_enable_derived_saops=true, yb_max_saop_merge_streams=64.",
-		buildNativeResolutionRecommendation("bucket-based indexes", constants.MATURITY_UNSUPPORTED, ybversion.V2024_2_0_0, ">=2025.2.1.0 (2025.2 series) (TP)", flags))
-
-	// No resolution name, or unsupported with no later versions -> empty.
-	assert.Empty(t, buildNativeResolutionRecommendation("", constants.MATURITY_TP, ybversion.V2025_2_1_0, "", flags))
-	assert.Empty(t, buildNativeResolutionRecommendation("bucket-based indexes", constants.MATURITY_UNSUPPORTED, ybversion.V2024_2_0_0, "", flags))
-}
-
-// withTargetDbVersion sets the package-level targetDbVersion for the duration of fn.
-func withTargetDbVersion(v *ybversion.YBVersion, fn func()) {
-	saved := targetDbVersion
-	targetDbVersion = v
-	defer func() { targetDbVersion = saved }()
-	fn()
-}
-
-func TestConvertIssueInstanceAnnotatesExperimentalMaturity(t *testing.T) {
-	// Feature that is Tech Preview in the 2025.2 target version, behind a flag.
-	qi := queryissue.QueryIssue{
-		Issue: issue.Issue{
-			Type:          "SOME_TP_FEATURE",
-			Name:          "Some TP Feature",
-			Description:   "Base description.",
-			Impact:        constants.IMPACT_LEVEL_1,
-			EnablingFlags: []string{"yb_enable_some_tp_feature=true"},
-			MinimumVersionsFixedInTP: map[string]*ybversion.YBVersion{
-				ybversion.SERIES_2025_2: ybversion.V2025_2_0_0,
-			},
-		},
-		ObjectType: "TABLE",
-		ObjectName: "public.t",
-	}
-
-	withTargetDbVersion(ybversion.V2025_2_0_0, func() {
-		analyzeIssue := convertIssueInstanceToAnalyzeIssue(qi, "", false, false)
-		// The converter appends exactly the experimental-maturity annotation to the base description.
-		expected := utils.JoinSentences("Base description.",
-			buildExperimentalMaturityAnnotation(constants.MATURITY_TP, ybversion.V2025_2_0_0, []string{"yb_enable_some_tp_feature=true"}))
-		assert.Equal(t, expected, analyzeIssue.Reason)
-	})
-}
-
-func TestConvertIssueInstancePerfRecommendationByTargetVersion(t *testing.T) {
-	// A performance optimization (hotspot) whose native resolution (bucket-based indexes) is
-	// Tech Preview in 2025.2.1. The resolution name is seeded in InternalDetails; the converter
-	// appends a version-aware recommendation to the Description.
-	newQi := func() queryissue.QueryIssue {
-		return queryissue.QueryIssue{
-			Issue: issue.Issue{
-				Type:        queryissue.HOTSPOTS_ON_DATE_INDEX,
-				Name:        "Hotspots with range sharded on date datatype index",
-				Description: "Base perf description.",
-				Impact:      constants.IMPACT_LEVEL_1,
-				MinimumVersionsFixedInTP: map[string]*ybversion.YBVersion{
-					ybversion.SERIES_2025_2: ybversion.V2025_2_1_0,
-				},
-				EnablingFlags: []string{"yb_enable_derived_saops=true", "yb_max_saop_merge_streams=64"},
-			},
-			ObjectType:      "INDEX",
-			ObjectName:      "idx_x ON public.t",
-			InternalDetails: map[string]interface{}{queryissue.RECOMMENDED_RESOLUTION: "bucket-based indexes"},
-		}
-	}
-
-	flags := []string{"yb_enable_derived_saops=true", "yb_max_saop_merge_streams=64"}
-	// Pinned literal: this is what the converter must derive from the issue's maps.
-	supportedVersions := ">=2025.2.1.0 (2025.2 series) (TP)"
-
-	// Resolution available (TP) in the target version: converter routes the perf issue through the
-	// native-resolution recommendation (not the experimental annotation) and appends it to Reason.
-	withTargetDbVersion(ybversion.V2025_2_1_0, func() {
-		analyzeIssue := convertIssueInstanceToAnalyzeIssue(newQi(), "", false, false)
-		expected := utils.JoinSentences("Base perf description.",
-			buildNativeResolutionRecommendation("bucket-based indexes", constants.MATURITY_TP, ybversion.V2025_2_1_0, supportedVersions, flags))
-		assert.Equal(t, expected, analyzeIssue.Reason)
-	})
-
-	// Older target: resolution not in target yet, recommendation lists where it lands.
-	withTargetDbVersion(ybversion.V2024_2_0_0, func() {
-		analyzeIssue := convertIssueInstanceToAnalyzeIssue(newQi(), "", false, false)
-		expected := utils.JoinSentences("Base perf description.",
-			buildNativeResolutionRecommendation("bucket-based indexes", constants.MATURITY_UNSUPPORTED, ybversion.V2024_2_0_0, supportedVersions, flags))
-		assert.Equal(t, expected, analyzeIssue.Reason)
-	})
 }
