+++ yb-voyager/src/query/queryissue/parser_issue_detector.go
@@ -19,6 +19,7 @@ package queryissue
 import (
 	"fmt"
 	"slices"
+	"sort"
 	"strings"
 
 	goerrors "github.com/go-errors/errors"
@@ -457,6 +458,11 @@ func (p *ParserIssueDetector) getIssuesNotFixedInTargetDbVersion(issues []QueryI
 			return nil, fmt.Errorf("checking maturity of issue %v in target version: %w", i, err)
 		}
 		if maturity != constants.MATURITY_GA {
+			// Surface the feature's maturity in the target version in the issue description, so every
+			// consumer (assessment report, analyze-schema, query constructs) reports it consistently.
+			if guidance := CheckIssueSupportMaturityInTDBVersion(i, targetDbVersion); guidance != "" {
+				i.Description = utils.JoinSentences(i.Description, guidance)
+			}
 			filteredIssues = append(filteredIssues, i)
 		} else {
 			if i.Issue.Type == UNLOGGED_TABLES {
@@ -465,6 +471,118 @@ func (p *ParserIssueDetector) getIssuesNotFixedInTargetDbVersion(issues []QueryI
 		}
 	}
 	return filteredIssues, nil
+}
+
+// GetSupportedVersions merges the GA/EA/TP version maps into a single human-readable
+// string, e.g. ">=2025.1.0.0 (2025.1 series) (EA), >=2025.2.0.0 (2025.2 series)". GA entries
+// stay untagged (so the "unsupported now, GA later" case is unchanged); TP/EA entries carry a
+// tier tag. Exported because the assessment report template renders it for the "Supported In
+// (versions)" row.
+func GetSupportedVersions(gaMap, eaMap, tpMap map[string]*ybversion.YBVersion) string {
+	var entries []string
+	appendEntries := func(versionsFixedIn map[string]*ybversion.YBVersion, tier string) {
+		for series, minVersionFixedIn := range versionsFixedIn {
+			if minVersionFixedIn == nil {
+				continue
+			}
+			if tier == constants.MATURITY_GA {
+				entries = append(entries, fmt.Sprintf(">=%s (%s series)", minVersionFixedIn.String(), series))
+			} else {
+				entries = append(entries, fmt.Sprintf(">=%s (%s series) (%s)", minVersionFixedIn.String(), series, tier))
+			}
+		}
+	}
+	appendEntries(gaMap, constants.MATURITY_GA)
+	appendEntries(eaMap, constants.MATURITY_EA)
+	appendEntries(tpMap, constants.MATURITY_TP)
+	sort.Strings(entries)
+	return strings.Join(entries, ", ")
+}
+
+// maturityTierName returns the display name + caveat for an experimental maturity tier.
+func maturityTierName(maturity string) (string, string) {
+	switch maturity {
+	case constants.MATURITY_TP:
+		return "Tech Preview (TP)", constants.TP_MATURITY_CAVEAT
+	case constants.MATURITY_EA:
+		return "Early Access (EA)", constants.EA_MATURITY_CAVEAT
+	default:
+		return "", ""
+	}
+}
+
+// buildExperimentalMaturityAnnotation returns a sentence (for an unsupported feature) describing
+// that the feature is available only as Tech Preview / Early Access in the target version, with
+// the flags needed to enable it.
+func buildExperimentalMaturityAnnotation(maturity string, targetDbVersion *ybversion.YBVersion, enablingFlags []string) string {
+	tierName, caveat := maturityTierName(maturity)
+	if tierName == "" {
+		return ""
+	}
+	annotation := fmt.Sprintf("This feature is available as %s in the target version (%s) — %s, and is not enabled by default.",
+		tierName, targetDbVersion.String(), caveat)
+	if len(enablingFlags) > 0 {
+		annotation += fmt.Sprintf(" Enable with the flag(s): %s.", strings.Join(enablingFlags, ", "))
+	}
+	return annotation
+}
+
+// buildNativeResolutionRecommendation returns a version-aware recommendation for a performance
+// optimization (e.g. "bucket-based indexes"). If the resolution is Tech Preview / Early Access in
+// the target version it reads "available as <tier> in the target version"; otherwise it lists where
+// it becomes available (supportedVersions).
+func buildNativeResolutionRecommendation(resolution, maturity string, targetDbVersion *ybversion.YBVersion, supportedVersions string, enablingFlags []string) string {
+	if resolution == "" {
+		return ""
+	}
+	var availability string
+	if tierName, caveat := maturityTierName(maturity); tierName != "" {
+		availability = fmt.Sprintf("available as %s in the target version (%s) — %s, and is not enabled by default",
+			tierName, targetDbVersion.String(), caveat)
+	} else if supportedVersions != "" {
+		availability = fmt.Sprintf("available in %s", supportedVersions)
+	} else {
+		return ""
+	}
+	rec := fmt.Sprintf("Consider using %s — %s.", resolution, availability)
+	if len(enablingFlags) > 0 {
+		rec += fmt.Sprintf(" Enable with the flag(s): %s.", strings.Join(enablingFlags, ", "))
+	}
+	return rec
+}
+
+// CheckIssueSupportMaturityInTDBVersion returns a human-readable note describing the feature's
+// maturity in the given target YugabyteDB version.
+func CheckIssueSupportMaturityInTDBVersion(issueInstance QueryIssue, targetDbVersion *ybversion.YBVersion) string {
+	if targetDbVersion == nil {
+		return ""
+	}
+
+	maturity, err := issueInstance.GetMaturityInTarget(targetDbVersion)
+	if err != nil {
+		log.Warnf("checking maturity of issue %q in target version: %v", issueInstance.Type, err)
+		return ""
+	}
+
+	// Performance optimizations: recommend the native resolution (e.g. bucket-based indexes),
+	// stating its maturity/availability in the target version and the flags to enable it.
+	if slices.Contains(PerformanceOptimizationIssues, issueInstance.Type) {
+		resolution, ok := issueInstance.InternalDetails[RECOMMENDED_RESOLUTION].(string)
+		if !ok {
+			return ""
+		}
+		supportedVersions := GetSupportedVersions(issueInstance.MinimumVersionsFixedIn, issueInstance.MinimumVersionsFixedInEA, issueInstance.MinimumVersionsFixedInTP)
+		return buildNativeResolutionRecommendation(resolution, maturity, targetDbVersion, supportedVersions, issueInstance.EnablingFlags)
+	}
+
+	// Unsupported features that are Tech Preview / Early Access in the target version:
+	// explain they are experimental and how to enable them.
+	if maturity == constants.MATURITY_TP || maturity == constants.MATURITY_EA {
+		return buildExperimentalMaturityAnnotation(maturity, targetDbVersion, issueInstance.EnablingFlags)
+	}
+
+	// GA in the target (already reported as fixed) or unsupported with no maturity data: nothing to add.
+	return ""
 }
 
 func (p *ParserIssueDetector) GetAllPLPGSQLIssues(query string, targetDbVersion *ybversion.YBVersion) ([]QueryIssue, error) {
